refactor: String 에서 NumberBall 변경 역할 Hitter 내부로 이동
BaseBallGame의 결합도를 줄이기 위해 기능을 이동시켜 NumbersBall에 대한 결합 제거
