+++ src/main/java/baseball/controller/BaseBallGame.java
@@ -3,10 +3,7 @@ package baseball.controller;
 import baseball.domain.Hitter;
 import baseball.domain.Pitcher;
 import baseball.domain.Result;
-import baseball.domain.numberBalls.NumberBalls;
 import baseball.view.BaseBallView;
-import java.util.Arrays;
-import java.util.stream.Collectors;
 
 
 public class BaseBallGame {
@@ -14,7 +11,6 @@ public class BaseBallGame {
     private static final String CONTINUE_GAME_SIG = "1";
     private static final String END_GAME_SIG = "2";
     private static final String WRONG_INPUT_ERROR_MESSAGE = "[ERROR] 잘못된 입력입니다.";
-    private static final String WRONG_INTEGER_ERROR_MESSAGE = "[ERROR] 숫자만 입력가능합니다.";
     private static final BaseBallView baseBallView  = BaseBallView.create();
 
     private static Pitcher pitcher;
@@ -54,23 +50,7 @@ public class BaseBallGame {
 
     private static void changeHitter() {
         String numberString = baseBallView.numbersInputView();
-        hitter = Hitter.create(convertStringToNumbers(numberString));
-    }
-
-    private static NumberBalls convertStringToNumbers(String numberString) {
-        validateIsInteger(numberString);
-
-        return NumberBalls.create(Arrays.stream(numberString.split(""))
-                .map(Integer::valueOf)
-                .collect(Collectors.toList()));
-    }
-
-    private static void validateIsInteger(String input) {
-        try {
-            Integer.valueOf(input);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(WRONG_INTEGER_ERROR_MESSAGE);
-        }
+        hitter = Hitter.create(numberString);
     }
 
     private static void checkContinueGame() {
